# accelerate line iterator