@cessen 2024-04-08
So, first off, I don't plan to accept PRs that introduce additional dependencies.  If anything, I want to reduce dependencies (e.g. remove `smallvec`).  I apologize for any wasted effort due to that.  I'll add that to the contributing section of the readme to make it clear for future contributors.

I don't think this PR is completely wasted effort, however.

The lack of optimization on reverse line iteration is already known, but it's good to see some hard numbers on how much faster it could be with SIMD.  The way I'd like to go about optimizing that case is to add functions to the [`str_indices`](https://github.com/cessen/str_indices) crate that handle reverse (from-the-end) index conversion.  I just haven't been motivated enough myself to tackle that yet.

> The forward search does use simd (line_to_byte_idx) but these functions from `str_indices` are more optimized for counting/larger counts and don't do quite as well in the latency department (when line_idx=1)

This is a great thing to identify, and like the reverse iteration case I'd like to see this tackled in [`str_indices`](https://github.com/cessen/str_indices) rather than bringing in another dependency.  We can probably add a special case in `line_to_byte_idx` for `line_idx=1` to do something faster, since that's a common case that's likely worth optimizing for.